infrastructure/containers/blockchain/requirements.chain-to-pay.txt
# Chain-to-pay gateway — wheels in Dockerfile.chain-to-pay (aligned with node-system-gateway)
pyyaml==6.0.1
